+++ portfolio/portfolio.html
@@ -6,10 +6,9 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <title>Ryszard Klimczewski - Technical Artist</title>
         <link rel="icon" type="image/svg+xml" href="/ryskdigital_logo.svg" />
-        <script type="module" src="main.js"></script>
+        <script type="module" src="portfolio.js"></script>
     </head>
     <body>
-        <div class="background-container | homepage-1"></div>
         <div class="navigation-bar | flex-column size-padding-l">
             <header class="size-margin-bottom-xl">
                 <div class="flex">
@@ -45,10 +44,10 @@
             <nav class="flex-grow size-font-xl" data-menu-state="closed">
                 <ul role="list" class="flex-column gap">
                     <li>
-                        <a data-current-page="true" href="#">Home</a>
+                        <a href="/index.html">Home</a>
                     </li>
                     <li>
-                        <a href="/portfolio/portfolio.html">Portfolio</a>
+                        <a data-current-page="true" href="#">Portfolio</a>
                     </li>
                     <li>
                         <a href="#">Experience</a>
