+++ portfolio/portfolio.js
@@ -1,0 +1,6 @@
+import "/scss/main.scss";
+import { toggleMenu, animateMenuIcon } from "/utilities.js";
+
+const nav = document.querySelector(".mobile-menu-button");
+nav.addEventListener("click", toggleMenu);
+nav.addEventListener("click", animateMenuIcon);
