+++ vite.config.js
@@ -6,6 +6,7 @@ export default defineConfig({
         rollupOptions: {
             input: {
                 main: resolve(__dirname, "index.html"),
+                portfolio: resolve(__dirname, "portfolio/portfolio.html"),
             },
         },
     },
